+++ src/core/assets/cleanup.json
@@ -1,1 +1,1 @@
-[{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com","method":"get","secrets":{}},{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com/elements/api-v2/instances/123","method":"delete","secrets":{"userSecret":"fake","orgSecret":"fake"}},{"url":"http://google.com","method":"get","secrets":{}}]
+[]
